Solved a bug in solver: the conversion from double to char was too naive, now a threshold approach has been adopted
Wrote a script to generate instances similar to the one in the TSP Partition Heuristic library.

Wrote a script that benches the time to process instances of various sizes, in order to understand the limits in which the threshold can vary, and automated the build process.

Added some sample instances too.